requirements.txt
flask
psycopg2-binary
werkzeug
cloudinary
